Add LICENSE file
The original compressed archive with the source code of the library
didn't include a license file. However, the author has expressed that
the library is licensed under the MIT license.

I've taken a look at the codebase to know the dates of the release and
modifications of the library code.

https://web.archive.org/web/20200918104259/https://www.coranac.com/projects/tonc/#comment-171904